backend/requirements.txt
fastapi
uvicorn
websockets
telethon
aiosqlite
unidecode
cryptg
python-dotenv